Handled the bugs for browse page and revoked some older code
