+++ static/js/edges.js
@@ -26,6 +26,7 @@ import {
 } from "./profile/peopleFrags.js";
 import { _renderWorkProfile, _renderWorkSidebar } from "./profile/workFrag.js";
 import { searchQueryObj } from "./search.js";
+import { getLabel } from "./helper/getFieldLabls.js";
 
 let emlo = {
   active: {},
@@ -49,7 +50,7 @@ let emlo = {
       console.warn(
         "Template is missing we are using the default template for edges"
       );
-      // this.template = new edges.templates.bs3.Facetview();
+      this.template = new edges.templates.bs3.Facetview();
     }
 
     if (this.components.length <= 0) {
@@ -456,6 +457,643 @@ emlo.DropDownRenderer = class extends edges.Renderer {
     return val;
   }
 };
+
+emlo.ResultTable = class extends edges.Component {
+  constructor(params) {
+    super(params);
+
+    // the secondary results to get the data from, if not using the primary
+    this.secondaryResults = edges.util.getParam(
+      params,
+      "secondaryResults",
+      false
+    );
+
+    // filter function that can be used to trim down the result set
+    this.filter = edges.util.getParam(params, "filter", false);
+
+    // a sort function that can be used to organise the results
+    this.sort = edges.util.getParam(params, "sort", false);
+
+    // the maximum number of results to be stored
+    this.limit = edges.util.getParam(params, "limit", false);
+
+    this.infiniteScroll = edges.util.getParam(params, "infiniteScroll", false);
+
+    this.infiniteScrollPageSize = edges.util.getParam(
+      params,
+      "infiniteScrollPageSize",
+      10
+    );
+
+    this.updateHeader = edges.util.getParam(params, "updateHeader", false);
+    this.headerSelector = edges.util.getParam(
+      params,
+      "headerSelector",
+      "header"
+    );
+
+    //////////////////////////////////////
+    // variables for tracking internal state
+
+    // the results retrieved from ES.  If this is "false" this means that no synchronise
+    // has been called on this object, which in turn means that initial searching is still
+    // going on.  Once initialised this will be a list (which may in turn be empty, meaning
+    // that no results were found)
+    this.results = false;
+
+    this.infiniteScrollQuery = false;
+
+    this.hitCount = 0;
+    this.highlighting = {};
+  }
+
+  synchronise() {
+    // reset the state of the internal variables
+    this.results = [];
+    this.infiniteScrollQuery = false;
+    this.hitCount = 0;
+
+    var source = this.edge.result;
+    if (this.secondaryResults !== false) {
+      source = this.edge.secondaryResults[this.secondaryResults];
+    }
+
+    // if there are no sources to pull results from, leave us with an empty
+    // result set
+    if (!source) {
+      return;
+    }
+
+    if (source && source.data && source.data.hasOwnProperty("highlighting")) {
+      this.highlighting = source.data["highlighting"];
+    }
+
+    // first filter the results
+    var results = source.results();
+    this._appendResults({ results: results });
+
+    // record the hit count for later use
+    this.hitCount = source.total();
+
+    if (this.updateHeader) {
+      this._updateHeader();
+    }
+
+    // --- REMOVE MODAL AFTER OPERATION ---
+    setTimeout(() => {
+      let doc = document.getElementById("remove-message");
+      if (doc) {
+        doc.style.display = "none";
+      }
+    }, 500); // remove after 0.5s for a smooth UX
+  }
+
+  _updateHeader() {
+    let currentDoc = document.getElementById(this.headerSelector);
+    try {
+      if (!currentDoc) {
+        return;
+      }
+
+      // Check if the fetching process is active
+      if (this.results === false) {
+        currentDoc.innerHTML = "Loading results...";
+        return;
+      }
+
+      if (this.hitCount > 0) {
+        let refineSearch = document.getElementById("refine_search");
+        refineSearch.style.display = "inline";
+      }
+
+      // Check if results are fetched correctly
+      if (this.hitCount && this.hitCount >= 0) {
+        if (this.hitCount > 50) {
+          currentDoc.innerHTML = `${this.hitCount} results (50 results per page)`;
+        } else {
+          currentDoc.innerHTML = `${this.hitCount} ${
+            this.hitCount > 1 ? "results" : "result"
+          }`;
+        }
+      } else {
+        // Fallback message when results are not fetched
+        currentDoc.innerHTML = "";
+      }
+    } catch (err) {
+      console.error(err);
+      currentDoc.innerHTML = "";
+    }
+  }
+
+  _appendResults(params) {
+    var results = params.results;
+
+    if (this.filter) {
+      results = this.filter({ results: results });
+    }
+
+    if (this.sort) {
+      results.sort(this.sort);
+    }
+
+    if (this.limit !== false) {
+      results = results.slice(0, this.limit);
+    }
+
+    this.results = this.results.concat(results);
+  }
+
+  infiniteScrollNextPage(params) {
+    var callback = params.callback;
+
+    // if we have exhausted the result set, don't try to get the next page
+    if (this.results.length >= this.hitCount) {
+      return;
+    }
+
+    if (!this.infiniteScrollQuery) {
+      this.infiniteScrollQuery = this.edge.cloneQuery();
+      this.infiniteScrollQuery.clearAggregations();
+    }
+
+    // move the from/size parameters to get us the next page
+    var currentSize = this.infiniteScrollQuery.getSize();
+    var currentFrom = this.infiniteScrollQuery.getFrom();
+    if (currentFrom === false) {
+      currentFrom = 0;
+    }
+    this.infiniteScrollQuery.from = currentFrom + currentSize;
+    this.infiniteScrollQuery.size = this.infiniteScrollPageSize;
+
+    var successCallback = edges.util.objClosure(
+      this,
+      "infiniteScrollSuccess",
+      ["result"],
+      { callback: callback }
+    );
+    var errorCallback = edges.util.objClosure(
+      this,
+      "infiniteScrollError",
+      false,
+      { callback: callback }
+    );
+
+    this.edge.queryAdapter.doQuery({
+      edge: this.edge,
+      query: this.infiniteScrollQuery,
+      success: successCallback,
+      error: errorCallback,
+    });
+  }
+
+  infiniteScrollSuccess(params) {
+    var results = params.result.results();
+    this._appendResults({ results: results });
+    params.callback();
+  }
+
+  infiniteScrollError(params) {
+    alert("error");
+    params.callback();
+  }
+};
+emlo.ResultTableRenderer = class extends edges.Renderer {
+  constructor(params) {
+    super(params);
+
+    // parameters that can be passed in
+    this.noResultsText = edges.util.getParam(
+      params,
+      "noResultsText",
+      "No results to display"
+    );
+
+    this.tableDisplay = edges.util.getParam(params, "tableDisplay", []);
+    this.showIndex = edges.util.getParam(params, "showIndex", true);
+    this.serialHeader = edges.util.getParam(params, "serialHeader", "#");
+    this.arrayValueJoin = edges.util.getParam(params, "arrayValueJoin", ", ");
+    this.omitFieldIfEmpty = edges.util.getParam(
+      params,
+      "omitFieldIfEmpty",
+      true
+    );
+
+    // New parameters for selection functionality
+    this.defaultSelected = edges.util.getParam(params, "defaultSelected", []);
+    this.showCheckbox = edges.util.getParam(params, "showCheckbox", false);
+    this.selectField = edges.util.getParam(params, "selectField", "uuid");
+    this.checkboxLimit = edges.util.getParam(params, "checkboxLimit", 10);
+    this.displayField = edges.util.getParam(params, "displayField", "");
+
+    this.selectedRows = new Set(); // Track selected rows by UUID
+    this.namespace = "edges-bs3-results-fields-by-table";
+
+    // Restore selection from URL on page load
+    this._restoreSelectionFromURL();
+  }
+
+  total() {
+    return this.component.hitCount;
+  }
+
+  draw() {
+    let frag = this.noResultsText;
+
+    if (this.component.results === false) {
+      frag = "Loading results... Please wait";
+    }
+    const resultHeader = document.getElementById("result-header-section");
+
+    const results = this.component.results;
+    if (results && results.length > 0) {
+      const recordClasses = edges.util.styleClasses(
+        this.namespace,
+        "record",
+        this.component.id
+      );
+
+      const urlParams = new URLSearchParams(window.location.search);
+      const shouldCall = urlParams.has("let_con"); // replace with actual param name
+
+      if (!shouldCall) {
+        this.tableDisplay = this.tableDisplay.filter(
+          (item) => item.field !== "let_con"
+        );
+      }
+
+      // create table headers
+      const headers = this.tableDisplay
+        .map((field) => `<th>${edges.util.escapeHtml(field.header)}</th>`)
+        .join("");
+      const headerRow = this.showIndex
+        ? `<tr>${this.showCheckbox ? "<th></th>" : ""}<th>${
+            this.serialHeader
+          }</th>${headers}</tr>`
+        : `<tr>${this.showCheckbox ? "<th></th>" : ""}${headers}</tr>`;
+
+      let rows = results
+        .map((result, index) => this._renderResult(result, index))
+        .join("");
+
+      frag = `
+            <table class="table table-bordered">
+                <thead>
+                    ${headerRow}
+                </thead>
+                <tbody>
+                    ${rows}
+                </tbody>
+            </table>
+        `;
+
+      if (resultHeader) {
+        resultHeader.style.display = "inline";
+      }
+
+      this._renderSideNav();
+    } else {
+      if (resultHeader) {
+        resultHeader.style.display = "none";
+      }
+    }
+
+    const containerClasses = edges.util.styleClasses(
+      this.namespace,
+      "container",
+      this.component.id
+    );
+
+    const container = `<div class="${containerClasses}">${frag}</div>`;
+
+    this.component.context.html(container);
+
+    // Attach event handlers for row selection
+    if (this.showCheckbox) {
+      this._attachRowSelectionHandlers();
+    }
+  }
+
+  _renderResult(res, index) {
+    const rowClasses = edges.util.styleClasses(
+      this.namespace,
+      "row",
+      this.component.id
+    );
+
+    const pageSize = this.component.infiniteScrollPageSize || 50;
+    let pageNumber = 1;
+
+    const paginationComponent = this.component.edge.components.find(
+      (comp) => comp.id === "top-pager"
+    );
+    if (paginationComponent && paginationComponent.page) {
+      pageNumber = paginationComponent.page;
+    }
+
+    // Retrieve existing query parameters from the current URL
+    const urlParams = new URLSearchParams(window.location.search);
+
+    // Initialize a query string for new or updated parameters
+    let queryParams = new URLSearchParams(urlParams);
+
+    const continuousIndex = (pageNumber - 1) * pageSize + index + 1;
+
+    const row = this.tableDisplay
+      .map((field) => {
+        let val = "";
+
+        if (field.field) {
+          val = this._getValue(field.field, res, val);
+        }
+        if (val) {
+          val = edges.util.escapeHtml(val);
+        }
+
+        if (field.valueFunction) {
+          val = field.valueFunction(
+            val,
+            res,
+            field.field,
+            this,
+            continuousIndex
+          );
+
+          if (field.header == "Repositories & Versions") {
+            return `<td id=repo-${continuousIndex}> </td>`;
+          } else {
+            return `<td>${val}</td>`;
+          }
+        }
+
+        if (!val && this.omitFieldIfEmpty) {
+          return "<td></td>";
+        }
+
+        if (field.type) {
+          if (field.type === "date") {
+            return `<td>${this._formatDate(val)}</td>`;
+          }
+
+          if (field.type === "pre") {
+            return `<td><pre>${val}</pre></td>`;
+          }
+
+          if (field.type === "link") {
+            let href = field.linkHref
+              ? this._getValue(field.linkHref, res, val)
+              : val;
+            let linkText = field.linkText || val;
+            let prefix = field.linkHrefPrefix || "";
+
+            if (prefix) {
+              let subPagen = res.hasOwnProperty("object_type")
+                ? res["object_type"]
+                : "";
+
+              if (subPagen != "") {
+                if (subPagen == "institution") {
+                  return `<td><a href="${prefix}/repository/${href}">${linkText}</a></td>`;
+                }
+                return `<td><a href="${prefix}/${subPagen}/${href}">${linkText}</a></td>`;
+              }
+
+              return `<td><a href="${prefix}/${href}">${linkText}</a></td>`;
+            }
+
+            return `<td><a href="${href}">${linkText}</a></td>`;
+          }
+
+          if (field.type === "multiple" && field.multipleFields) {
+            const multipleFieldDisplay = field.multipleFields
+              .map((item) => {
+                let value = "";
+
+                if (item.isSemiColon) {
+                  if (
+                    res &&
+                    res.hasOwnProperty(item.field) &&
+                    res[item.field]
+                  ) {
+                    value = res[item.field].split("\n").join("; ");
+                  }
+                } else {
+                  value = this._getValue(item.field, res, "");
+                }
+                return value ? `<div>${item.label}: ${value}</div>` : "";
+              })
+              .join("");
+            return `<td>${multipleFieldDisplay}</td>`;
+          }
+        }
+
+        return `<td>${val}</td>`;
+      })
+      .join("");
+
+    const data = this._getSelectField(res[this.selectField], this.selectField);
+    const isChecked = this.defaultSelected.includes(data) ? "checked" : "";
+    if (isChecked) {
+      this.selectedRows.add(data);
+    }
+
+    const checkboxCell = this.showCheckbox
+      ? `<td><input type="checkbox" class="select-row" data-uuid="${data}" data-display="${
+          res[this.displayField]
+        }" ${isChecked}></td>`
+      : "";
+
+    return this.showIndex
+      ? `<tr class="${rowClasses}">${checkboxCell}<td>${continuousIndex}</td>${row}</tr>`
+      : `<tr class="${rowClasses}">${checkboxCell}${row}</tr>`;
+  }
+
+  _getSelectField(data, selectField) {
+    if (selectField == "uuid") {
+      return data;
+    } else {
+      if (Array.isArray(data) && data.length > 0) {
+        return data.map((url) => url.split("/").pop()).join(", ");
+      }
+      return "";
+    }
+  }
+
+  _renderSideNav() {
+    const selectedItems = Array.from(this.selectedRows)
+      .slice(0, this.checkboxLimit)
+      .map((uuid) => {
+        const displayName = this._getDisplayName(uuid) || uuid;
+        return `
+          <input type="checkbox" class="side-nav-item" data-uuid="${uuid}" checked>
+          <label> 
+            <span> ${displayName} </span>
+          </label>
+          <br/>
+        `;
+      })
+      .join("");
+
+    const selectedList = document.getElementById("selected-items-list");
+    const sideNavDoc = document.getElementById("side-nav");
+    if (selectedList) {
+      selectedList.innerHTML = selectedItems;
+    }
+
+    if (sideNavDoc) {
+      if (this.selectedRows.size > 0) {
+        sideNavDoc.style.display = "inline";
+      } else {
+        sideNavDoc.style.display = "none";
+      }
+
+      // Update the URL
+      this._updateURL();
+
+      // Attach event handlers for side nav items
+      this._attachSideNavHandlers();
+    }
+  }
+
+  _attachRowSelectionHandlers() {
+    const context = this.component.context;
+    const renderer = this;
+
+    context.find(".select-row").on("change", function () {
+      const uuid = $(this).data("uuid");
+      const displayName = $(this).data("display");
+
+      // Check the current selection count
+      if ($(this).is(":checked")) {
+        if (renderer.selectedRows.size >= renderer.checkboxLimit) {
+          // Prevent additional selections if limit is reached
+          $(this).prop("checked", false);
+
+          // Display a message to the user
+          alert(`You can only select up to ${renderer.checkboxLimit} items.`);
+          return;
+        }
+
+        // Add the UUID to the selected set
+        renderer.selectedRows.add(uuid);
+      } else {
+        // Remove the UUID from the selected set if unchecked
+        renderer.selectedRows.delete(uuid);
+      }
+
+      // Update the side navigation and URL
+      renderer._renderSideNav();
+    });
+  }
+
+  _attachSideNavHandlers() {
+    const renderer = this;
+
+    // Attach event handlers to the side navigation items
+    const sideNavItems = document.querySelectorAll(".side-nav-item");
+
+    sideNavItems.forEach((item) => {
+      item.addEventListener("change", function () {
+        const uuid = this.getAttribute("data-uuid");
+
+        if (!this.checked) {
+          // Remove the UUID from the selected set
+          renderer.selectedRows.delete(uuid);
+
+          // Uncheck the corresponding checkbox in the table
+          const tableCheckbox = renderer.component.context.find(
+            `.select-row[data-uuid='${uuid}']`
+          );
+          if (tableCheckbox.length) {
+            tableCheckbox.prop("checked", false);
+          }
+
+          // Update the URL to reflect the change
+          renderer._updateURL();
+
+          renderer._renderSideNav();
+        }
+      });
+    });
+  }
+
+  _getValue(path, rec, def) {
+    if (def === undefined) {
+      def = false;
+    }
+    const bits = path.split(".");
+    let val = rec;
+    for (let i = 0; i < bits.length; i++) {
+      const field = bits[i];
+      if (field in val) {
+        val = val[field];
+      } else {
+        return def;
+      }
+    }
+    if ($.isArray(val)) {
+      val = val.join(this.arrayValueJoin);
+    } else if ($.isPlainObject(val)) {
+      val = def;
+    }
+    return val;
+  }
+
+  _getDisplayName(uuid) {
+    const results = this.component.results || [];
+    const result = results.find((res) => res.uuid === uuid);
+    return result ? result[this.displayField] : null;
+  }
+
+  _updateURL() {
+    const uuids = Array.from(this.selectedRows).join(",");
+    const url = new URL(window.location);
+    if (uuids) {
+      url.searchParams.set("uuids", uuids);
+    } else {
+      url.searchParams.delete("uuids");
+    }
+
+    window.history.replaceState({}, "", url);
+  }
+
+  _restoreSelectionFromURL() {
+    const urlParams = new URLSearchParams(window.location.search);
+    const uuids = urlParams.get("uuids");
+    if (uuids) {
+      this.defaultSelected = uuids.split(",");
+    }
+  }
+
+  _formatDate(timestamp) {
+    const date = new Date(timestamp);
+    if (isNaN(date.getTime())) {
+      return "";
+    }
+
+    const months = [
+      "January",
+      "February",
+      "March",
+      "April",
+      "May",
+      "June",
+      "July",
+      "August",
+      "September",
+      "October",
+      "November",
+      "December",
+    ];
+
+    const day = date.getDate().toString().padStart(2, "0");
+    const month = months[date.getMonth()];
+    const year = date.getFullYear();
+
+    return year === 9999 ? `${day} ${month}` : `${day} ${month} ${year}`;
+  }
+};
+
 emlo.Facet = class extends edges.components.RefiningANDTermSelector {
   constructor(params) {
     super(params);
@@ -1179,6 +1817,233 @@ emlo.FacetRenderer = class extends edges.Renderer {
     );
     // refresh the event binding
     edges.on(tooltipSelector, "click", this, "toggleTooltip");
+  }
+};
+
+emlo.SelectedFacetRenderer = class extends edges.Renderer {
+  constructor(params) {
+    super(params);
+
+    ///////////////////////////////////////
+    // parameters that can be passed in
+    this.title = edges.util.getParam(params, "title", "");
+    this.namespace = "emlo-selected-facet-view";
+  }
+
+  draw() {
+    let ts = this.component;
+
+    // Clear the context if no filters are active
+    if (ts.filters.length === 0) {
+      ts.context.html(
+        "<table class='facet'><tbody><tr><td>None</td></tr></tbody></table>"
+      );
+      return;
+    }
+
+    // Class selectors for styling
+    const facetClass = edges.util.styleClasses(
+      this.namespace,
+      "facet",
+      this.component.id
+    );
+    const headerClass = edges.util.styleClasses(
+      this.namespace,
+      "header",
+      this.component.id
+    );
+    const selectedClass = edges.util.styleClasses(
+      this.namespace,
+      "selected",
+      this.component.id
+    );
+    const resultClass = edges.util.styleClasses(
+      this.namespace,
+      "result",
+      this.component.id
+    );
+    const filterRemoveClass = edges.util.allClasses(
+      this.namespace,
+      "filter-remove",
+      this.component.id
+    );
+
+    // Build the selected filters display
+    let filterFrag = "";
+    ts.filters.forEach((filt) => {
+      filterFrag += `
+        <tr class="${resultClass}" style="vertical-align: middle;">
+          <td style="width:80px" class="capitalize-first-letter">
+          ${this._getSelectedFieldLabel(filt.field)}
+          </td>
+
+          <td class="${filterRemoveClass} selected-facets" data-key='${
+        filt.term
+      }' data-field='${filt.field}' data-render="selected">
+              <span data-val='${filt.field}'  style="width:100px">
+              ${this._getDisplayValue(filt.field, filt.display)}
+              </span>
+              <span style="margin-left:15px;widht:50px">
+                <img class="facet" src="../../static/img/minus-facet.png" style="height:15px;" />
+              </span>
+            </td>
+        </tr>
+      `;
+    });
+
+    let frag = `<div class="${facetClass}">
+                  <div class="${headerClass}">
+                    <h4>${this.title}</h4>
+                    <p id="remove-message"> Please wait, we are updating the results. </p>
+                  </div>
+                  <div class="${selectedClass}">
+                  <table class="facet">
+                    <tbody>
+                      ${filterFrag}
+                    </tbody>
+                  </table>
+                  </div>
+                </div>`;
+
+    // Render the HTML in the component context
+    ts.context.html(frag);
+
+    // Set event handler for removing selected filters
+    const filterRemoveSelector = edges.util.jsClassSelector(
+      this.namespace,
+      "filter-remove",
+      this.component.id
+    );
+    edges.on(filterRemoveSelector, "click", this, "removeFilter");
+  }
+
+  removeFilter(element) {
+    const term = element.getAttribute("data-key");
+    const field = element.getAttribute("data-field");
+
+    this.component.removeFilter(field, term);
+    this.draw(); // Redraw the component to reflect the changes
+  }
+
+  // PATCH: currently we do not have anything in edges that can help us with this.
+  _getDisplayValue(field, value) {
+    const colMap = {
+      "mail_origin-location": "locations",
+      "mail_destination-location": "locations",
+      "dcterms_references-location": "locations",
+      "frbr_creator-person": "people",
+      "mail_recipient-person": "people",
+      "dcterms_references-person": "people",
+    };
+
+    const displayFieldMapping = {
+      work: "Letter",
+      manifestation: "Document",
+      resource: "Related resource",
+      person: "Person or organization",
+      location: "Location",
+      image: "Image",
+      institution: "Repository",
+      comment: "Comment",
+    };
+
+    const validFields = [
+      "uuid_related",
+      "dcterms_references-location",
+      "mail_destination-location",
+      "mail_origin-location",
+      "frbr_creator-person",
+      "mail_recipient-person",
+      "dcterms_references-person",
+    ];
+
+    if (validFields.includes(field)) {
+      // Return a placeholder value immediately
+      const placeholder = "Loading...";
+
+      if (value && value.startsWith("*") && value.endsWith("*")) {
+        value = value.slice(1, -1);
+      }
+
+      if (
+        typeof value === "string" &&
+        value.startsWith('"http') &&
+        value.endsWith('"')
+      ) {
+        value = value.split("/").filter(Boolean).pop();
+        value = value.slice(0, -1);
+      }
+
+      let collectionName = "";
+
+      if (colMap.hasOwnProperty(field)) {
+        collectionName = colMap[field];
+      }
+
+      // Fetch names asynchronously
+      this._fetchNamesSelected(value, collectionName).then((names) => {
+        if (names) {
+          // Find all matching elements dynamically and update their content
+          document
+            .querySelectorAll(`[data-val="${edges.util.escapeHtml(field)}"]`)
+            .forEach((el) => {
+              el.innerHTML = `
+                ${edges.util.escapeHtml(names)}
+              `;
+            });
+        }
+      });
+
+      return placeholder;
+    } else {
+      if (value && value.startsWith('"') && value.endsWith('"')) {
+        value = value.slice(1, -1);
+      }
+
+      if (field == "object_type" && displayFieldMapping.hasOwnProperty(value)) {
+        return displayFieldMapping[value];
+      }
+
+      return value;
+    }
+  }
+
+  async _fetchNamesSelected(value, colName) {
+    let collectionName = "";
+    let fl = "browse";
+
+    if (colName) {
+      collectionName = colName;
+    } else {
+      const urlParams = new URLSearchParams(window.location.search);
+      const browsing = urlParams.get("browsing");
+
+      collectionName =
+        browsing && browsing != "organizations" ? `${browsing}` : `people`;
+    }
+
+    if (!collectionName) {
+      console.error("Collection name not found in the URL.");
+      return "";
+    }
+
+    const response = await fetch(
+      `/solr/${collectionName}/select?q=uuid:${value}&fl=${fl}&wt=json`
+    );
+    const data = await response.json();
+
+    // Extract and process `browse` values
+    const browseNames = data.response.docs.map((doc) => doc[fl]);
+
+    const browseNamesString = browseNames.join(", ");
+
+    return browseNamesString;
+  }
+
+  _getSelectedFieldLabel(field) {
+    const label = getLabel(field);
+
+    return label == "-" ? field : label;
   }
 };
