+++ templates/_edges_common_js.jinja2
@@ -7,5 +7,16 @@
 
 {# Loading edges components #}
 <script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/components/RefiningANDTermSelector.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/components/ResultsDisplay.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/components/FullSearchController.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/components/Pager.js') }}"></script>
+
+{# Loading edges default template #}
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/templates/bs3/Facetview.js') }}"></script>
 
 {# Loading edges renderer #}
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/renderers/bs3/RefiningANDTermSelector.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/renderers/bs3/ResultsFieldsByRow.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/renderers/bs3/FullSearchController.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/renderers/bs3/Pager.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='libs/edges/src/renderers/bs5/SearchBox.js') }}"></script>
